+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/MagicController.java
@@ -302,7 +302,8 @@ public class MagicController implements MageController {
             "target_armor", "target_luck", "target_knockback_resistance",
             "target_location_x", "target_location_y", "target_location_z",
             "target_mana", "target_mana_max",
-            "target_air", "target_air_max", "target_hunger", "target_fall_distance"
+            "target_air", "target_air_max", "target_hunger", "target_fall_distance",
+            "target_humidity", "target_temperature"
     );
     private final Set<String> builtinAttributes = ImmutableSet.of(
             "epoch",